requirements.txt
streamlit==1.37.0
streamlit-cropper==0.2.2
Pillow==10.4.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
easyocr==1.7.1
requests==2.32.3
google-genai>=0.3.0